+++ src/main/java/com/metrosys/rewards/Reward.java
@@ -5,10 +5,10 @@ package com.metrosys.rewards;
 public class Reward {
 
     private String month;
-    private short total;
-    private short[] values;
+    private int total;
+    private int[] values;
 
-    public Reward(String month, short[] values) {
+    public Reward(String month, int[] values) {
         this.month = month;
         this.values = values;
         this.total = getTotal(); // calculate the total on the object creation
@@ -18,14 +18,14 @@ public class Reward {
         return month;
     }
 
-    public short getTotal(){
+    public int getTotal(){
         // Reward rule calculation:
         // 2 points for every dollar spent over $100
         final short twoPointLimit = 100;
         // 1 point for every dollar spent over $50
         final short onePointLimit = 50;
 
-        short total = 0;
+        int total = 0;
         for (int i = 0; i < values.length; i++) {
             if (values[i] > twoPointLimit ) {
                 total += 2 * (values[i] - twoPointLimit) + (1 * onePointLimit);
